App/requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langchain-text-splitters
langchain-core
chromadb
sentence-transformers
pypdf
python-dotenv
huggingface-hub